Thistlewaite and Campbell (1960) Regression Discontinuity Data

This contains the data used in Thistlewaite and Campbell (1960), the first regression-discontinuity design paper. It studied the effect of merit awards on future academic outcomes: career aspirations, enrollment in Ph.D. or M.D. programs, and intellectualism.

CampbellThistlewaite1960

Format

A data frame with 22 rows and 10 columns.

name 	 type 	 description

certificate_of_merit 	 "integer" 	 Received a certificate of merit. From Table 1.

test_score 	 "character" 	 Aptitude score interval, relative to the cutoff score of the student's state. From Table 1.

merit_scholars 	 "integer" 	 Number of merit scholars in the interval.

number_in_sample 	 "integer" 	 The number of students in the sample. From Table 1.

number_respondents 	 "integer" 	 The number of students responding to the survey. From Table 1.

scholarship 	 "numeric" 	 Percent of individuals winning any scholarship. From Fig 2.

scholarship_150 	 "numeric" 	 Percent of individuals receiving a scholarship of \$150 or greater. From Fig. 2.

graduate_study 	 "numeric" 	 Percent of individuals reporting that they were planning to do three or more years of graduate study (PhD or M.D.). From Fig 3.

teacher_researcher 	 "numeric" 	 Percent of individuals reporting that they were planning to be a college teacher or scientific researcher. From Fig 3.

intellectualism 	 "numeric" 	 Mean score on an attitude towards intellectualism scale.

Source

Thistlewaite, Donald L. and Donald T. Campbell. (1960) "Regression-Discontinuity Analysis: An Alternative to the Ex Post Facto Experiment." Journal of Educational Psychology. 51(6), December 1960. pp. 309-317.

Details

Their sample consists near winners in the 1957 Merit program. It consists of 5,126 students who received the Certificates of Merit (C of M) and 2,848 students who received a letter of commendation. Students received the Certificate of Merit primarily based on the basis of the CEEB Scholarship Qualifying Test (SQT). The sample only includes those within 10 points of the cut-off point for their state. These students were sent a questionnaire which provides the response variables for this analysis.

The values of scholarship, scholarship_150, number_respondents, teacher_researcher, and intellectualism come from extracting data from the figures in the paper.