mphPlanet Money Cow Weight Guesses

Data collected by the NPR podcast Planet Money to the question, "How much does this cow weigh?". This is a replication of the Galton paper "Vox Populi" which cites guesses of cow's weight from a county fair to illustrate the "wisdom of crowds".

PlanetMoneyCowGuesses

Format

A data frame with 17184 rows and 2 columns.

name 	 type 	 description

guess 	 "numeric" 	 Response to the question, "How much does this cow weigh?". The weights are in pounds.

error 	 "numeric" 	 Difference (in pounds) between the guess and the weight of 1,355 pounds.

Source

Planet Money Cow Data

References

Quoctrung Bui, "17,205 People Guessed The Weight Of A Cow. Here's How They Did.", Planet Money, August 7, 2015. http://www.npr.org/sections/money/2015/08/07/429720443/17-205-people-guessed-the-weight-of-a-cow-heres-how-they-did.

Jacob Goldstein, "How Much Does This Cow Weigh?," Planet Money, July 17, 2015.

See also

VoxPopuli for the data from original Galton paper which inspired this data collection.