Votes in a cow weight-judging competition

Distribution of 787 guesses in a cow weight-judging competition. These data was used by Galton to in a series of letters to Nature to illustrate the wisdom of crowds.

VoxPopuli

Format

A data frame with 19 rows and 3 columns.

name 	 type 	 description

quantile 	 "numeric" 	 Quantile of the distribution of guesses.

estimate 	 "numeric" 	 Value of the guess (lbs.)

normal 	 "numeric" 	 Quantile values of a normal distribution with a median of 1,207 lbs. and a probable error (half the interquartile range) of 37 lbs (standard deviation of 25 lbs).

Source

Galton, Francis. 1907. "Vox Populi." Nature. doi:10.1038/075414a0

Details

The actual weight of the cow was 1,198 pounds.

References

Galton, Francis. 1907. "One Value, One Vote." Nature. doi:10.1038/075414a0

Galton, Francis. 1907. "Ballot Box." Nature doi:10.1038/075509e0

Hooke, R. H. 1907. "Mean or Median." Nature. doi:10.1038/075487a0

Yule, G. Udny. 1907. "Mean or Median." Nature. doi:10.1038/075534b0

Kenneth W. Wallis. 2014. "Revisiting Francis Galton's Forecasting Competition" https://arxiv.org/pdf/1410.3989.pdf

Levy, D. M. and Peart, S. (2002). Galton's two papers on voting as robust estimation. Public Choice 113 357–365.