Prices of Bordeaux Vintages

Prices of Vintages of Bordeaux Wines in New York auctions. The price is an index based on the wines of multiple chateaus, from a regression of the logarithm price from several thousand auction sales with variables for chateau and vintage years. These data include vintages 1952 to 1980, with missing price values for 1954 and 1956 because they were rarely sold.

bordeaux

Format

A data frame with 38 rows and 6 columns.

name 	 type 	 description

vint 	 "integer" 	 Vintage year

price 	 "numeric" 	 Price of the vintage, as an index (1961=100).

wrain 	 "integer" 	 Winter (October to March) rain in ML.

degrees 	 "numeric" 	 Average temperature (degrees centigrade) April to September.

hrain 	 "integer" 	 Harvest (August and September) rain in ML

time_sv 	 "integer" 	 Time since vintage - relative to 1983

Source

http://www.liquidasset.com/winedata.html

References

Ashenfelter, O. and Ashmore, D. and Lalonde, R. (1995) "Bordeaux Wine Vintage Quality and the Weather." Chance. doi:10.1080/09332480.1995.10542468

Ashenfelter, O. (2008). "Predicting the Quality and Prices of Bordeaux Wine." The Economic Journal doi:10.1111/j.1468-0297.2008.02148.x