Borough's 1580 data for the variation of the needle

William Borough's data for readings of the variation between magnetic and true north using the Sun's taken at different times of the day at the Limehouse lighthouse, not far from the Greenwich Meridian. These readings was published in A Discours of the Variation of the Campas, or Magneticall Needle.

borough

Format

A data frame with 9 rows and 4 columns.

name 	 type 	 description

elevation 	 "integer" 	 Elevation of the Sun in degrees

variation_am 	 "numeric" 	 Difference between the shadow of the Sun before noon and the compass north.

variation_pm 	 "numeric" 	 Difference between the shadow of the Sun after noon and the compass north.

variation_north 	 "numeric" 	 Variation between true and magnetic north.

Source

Stigler, Stephen M. The Seven Pillars of Statistical Wisdom, p. 20

References

Stigler, Stephen M. The Seven Pillars of Statistical Wisdom

See also

gellibrand and db1668 for other datasets relevant to the history of the mean.