King-Davenant Price Schedule

The King-Davenant price schedule is a small schedule of the the harvest sizes and prices of corn. It was published by Davenant in 1699 and also attributed to Gregory King. This schedule was analyzed by Jevons in 1871. It is likely that this schedule is only an approximation of the original data, since the the third difference is constant (Stigler 1999, p. 81). Jevons (1871) fit the schedule with a function of the form $$\frac{a}{{(x - b)}^{n}}$$. It assumed \(n = 2\) was assumed, and then solved for \(a = 0.824\) and \(b = 0.12\). See Stigler (1999) for details of how this was fit.

davenant

Format

A data frame with 6 rows and 3 columns.

name 	 type 	 description

quantity 	 "numeric" 	 Harvest size of corn.

price 	 "numeric" 	 Price of corn.

fitted 	 "numeric" 	 Jevons's (1871) fitted values

Source

Stigler, S. M. (1999) Statistics on the Table. p 81.

References

Jevons, W. S. (1871) Theory of Political Economy Stigler, S. M. (1999) Statistics on the Table. p 80-86.