Table of Magnetic Variation near Bristol on May 23, 1668

This is from a note published in the Transactions of the Royal Society in 1668. The editor, Henry Oldenburg, printed parts of a letter from "D.B." with five values of the magnetic variation near Bristol, England on May 23, 1668. This letter is notable in that he calculates the "mean of the true Variation", where "mean" refers to the arithmetic mean. It is referenced by Stigler as the first instance in which the term "mean" is used to refer to the arithmetical mean. D.B. reported a mean of 1 degree 27 minutes, rounding down from the arithmetic mean of 1 degree 28 minutes.

db1668

Format

A data frame with 5 rows and 4 columns.

name 	 type 	 description

sun_altitude 	 "numeric" 	 Sun's altitude in degrees

magnetic_azimuth 	 "numeric" 	 Magnetic azimuth in degrees

sun_azimuth 	 "numeric" 	 Sun's azimuth in degrees

variation 	 "numeric" 	 Variation in degrees westerly, between true north (sun azimuth) and magnetic north (magnetic azimuth)

Source

Stigler, Stephen M. The Seven Pillars of Statistical Wisdom

References

Stigler, Stephen M. The Seven Pillars of Statistical Wisdom, p. 26

See also

gellibrand and borough for other early datasets used to (almost) calculate means.