Price of Wheat in Delhi 1763-1836

Price of wheat in Delhi 1763-1836. This series was notoriously and spuriously related to sunspots (Wolf's numbers) by William S. Jevons in his last published paper. Jevons theorized that sunspots were related to commercial cycles and crises due to similarity in cycles of these time series.

delhiwheat

Format

A data frame with 74 rows and 4 columns.

name 	 type 	 description

year 	 "integer" 	 Year

price 	 "integer" 	 Price in rupees per 1,000 seers of wheat in Delhi

maximum 	 "logical" 	 Indicator marking a price maxima in a cycle

crisis 	 "logical" 	 Commercial crisis

Source

Jevons, W. Stanley. 1879. "Sun-Spots and Commercial Crises." Nature. doi:10.1038/019588a0.

References

Stigler, Stephen M. The Seven Pillars of Statistical Wisdom

Jevons, W. Stanley. 1882. "The Solar-Commercial Cycle." Nature. doi:10.1038/026226a0.

See also

nonsense for another example of spurious correlations.