French Lottery Winning Bets 1779-1833

The winning quaterne (four-number) bets for the French lottery, 1797 to 1833. In the French lottery (Loterie), a drawing consisted of drawing five balls without replacement, from a bin of balls numbered 1 to 90. Players could place bets by specifying a set of numbers, which had different names and payoffs based on the number specified.

loterie_bets

Format

A data frame with 327 rows and 4 columns.

name 	 type 	 description

X1 	 "integer" 	 First number

X2 	 "integer" 	 Second number

X3 	 "integer" 	 Third number

X4 	 "integer" 	 Fourth number

Source

http://www.stat.uchicago.edu/~stigler/French_Lottery_Data.html

Details

name 	 number correct 	 payoff

quine 	 five 	 1,000,000

quaterne 	 four 	 75,000

terne 	 three 	 5,000

ambe 	 two 	 270

References

Stephen M. Stigler. 2003. “Casanova, Bonaparte and the Loterie de France”, Journal de la Société Française de Statistique

Stigler, Stephen M. The Seven Pillars of Statistical Wisdom, p. 166-170.