Yule and Kendall's nonsense correlations

Data on the number of radio receiver licenses in the United Kingdom, 1927-1937 and the number of notified mental defectives per 10,000 in the United Kingdom were used by Udny Yule and Maurice Kendall in An Introduction to the Theory of Statistics to illustrate what they called "nonsense correlations". Tufte updated these variable with the number of letters in the first name of the US president for those years.

nonsense

Format

A data frame with 14 rows and 4 columns.

name 	 type 	 description

year 	 "integer" 	 Year

radio_licenses 	 "numeric" 	 Number of radio receiver licenses in the United Kingdom, 1927-1937

mental_defectives 	 "integer" 	 Number of notified "mental defectives" per 10,000 in the United Kingdom

president 	 "character" 	 First name of the U.S. President

Source

Tufte, Edward R. 1974. Data Analysis for Politics and Policy. Prentice Hall.

References

Tufte, Edward R. 1974. Data Analysis for Politics and Policy. Prentice Hall.

Yule, G. Udny, and M.G. Kendall. 1950. An Introduction to the Theory of Statistics. 14th ed. London:Charles Griffin. p. 315-16.