Pauperism in England, 1871-1891 (Poor Law Unions)

Data on pauperism in England, 1871-1881, used in Udny Yule's 1899 paper. This data frame contains Poor Law Union specific variables.

pauperism_plu

Format

A data frame with 599 rows and 7 columns.

name 	 type 	 description

ID 	 "numeric" 	 ID

Union 	 "character" 	 Union

Region 	 "character" 	 Region

County 	 "character" 	 County

Type 	 "character" 	 Yule's Rural/Urban Classification of Unions

BoothGroup 	 "character" 	 Booth's Classification of Unions

BoothDensity 	 "numeric" 	 Booth's Population Density

Details

See pauperism_year for further description of the data and the yearly pauperism statistics.

References

Yule, G. U. (1899) “An Investigation into the Causes of Changes in Pauperism in England, Chiefly During the Last Two Intercensal Decades (part I.)”. Journal of the Royal Statistical Society doi:10.2307/2979889.

Plewis, I. (2015) “Multiple regression, longitudinal data and welfare in the nineteenth century: reflections on Yule (1899)”, submitted to Journal of the Royal Statistical Society, Series A.

Plewis, I. and Schmidt, E. (2015) “Reconstructing G. Udny Yule’s 1899 pauperism dataset”, CMIST Report 2015-3