Pauperism in England, 1871-1891 (Yearly Data)

Data on pauperism used in the 1899 article, “An Investigation into the Causes of Changes in Pauperism in England”, by Udny Yule, which is the first use of multiple regression. The data contains Yearly data on pauperism, the proportion of paupers provided for at home vs. the work house, population, and population over 65 for the Poor Law Unions of England for the years: 1871, 1881, and 1891.

pauperism_year

Format

A data frame with 1797 rows and 38 columns.

name 	 type 	 description

ID 	 "numeric" 	 ID

year 	 "integer" 	 Year

adultchabI 	 "numeric" 	 Ratio of Adult to Child Able-Bodied In-Door Paupers

adultchabO 	 "numeric" 	 Ratio of Adult to Child Able-Bodied Out-Door Paupers

adultchnotabI 	 "numeric" 	 Ratio of Adult to Child Not Able-Bodied In-Door Paupers

adultchnotabO 	 "numeric" 	 Ratio of Adult to Child Not Able-Bodied Out-Door Paupers

ChABI 	 "numeric" 	 Child Able-Bodied In-Door Paupers

ChABO 	 "numeric" 	 Child Able-Bodied Out-Door Paupers

ChNotABI 	 "numeric" 	 Child Not Able-Bodied In-Door Paupers

ChNotABO 	 "numeric" 	 Child Not Able-Bodied Out-Door Paupers

F65 	 "numeric" 	 Females Aged 65+ in the Population

FABI 	 "numeric" 	 Female Able-Bodied In-Door Paupers

FABO 	 "numeric" 	 Female Able-Bodied Out-Door Paupers

FNotABI 	 "numeric" 	 Female Not Able-Bodied In-Door Paupers

FNotABO 	 "numeric" 	 Female Not Able-Bodied Out-Door Paupers

M65 	 "numeric" 	 Males Aged 65+ in the Population

MABI 	 "numeric" 	 Male Able-Bodied In-Door Paupers

MABO 	 "numeric" 	 Male Able-Bodied Out-Door Paupers

mfabI 	 "numeric" 	 Ratio of Male to Female Able-Bodied In-Door Paupers

mfabO 	 "numeric" 	 Ratio of Male to Female Able-Bodied Out-Door Paupers

mfnotabI 	 "numeric" 	 Ratio of Male to Female Not Able-Bodied In-Door Paupers

mfnotabO 	 "numeric" 	 Ratio of Male to Female Not Able-Bodied Out-Door Paupers

MNotABI 	 "numeric" 	 Male Not Able-Bodied In-Door Paupers

MNotABO 	 "numeric" 	 Male Not Able-Bodied Out-Door Paupers

outratio 	 "numeric" 	 Ratio of Out-Door Paupers to In-Door Paupers

outratiodiff 	 "numeric" 	 Ratio of outratio91 to outratio81

pauper 	 "numeric" 	 Proportion of Paupers in the Population

pauper2 	 "numeric" 	 Proportion of Paupers in the Population correcting for border changes.

paupratiodiff 	 "numeric" 	 Ratio of Pauperism Proportion to previous decade

Popn 	 "numeric" 	 Population (from Census Population Abstracts)

Popn2 	 "numeric" 	 Population allowing for boundary changes. This is the population number used by Yule.

PopnF 	 "numeric" 	 Female Population

PopnM 	 "numeric" 	 Male Population

popratiodiff 	 "numeric" 	 Ratio of Population to previous decade

PropF65 	 "numeric" 	 Proportion of females aged 65+ in the female population

PropM65 	 "numeric" 	 Proportion of males aged 65+ in the Male Population

Prop65 	 "numeric" 	 Proportion of the population Aged 65+

oldratiodiff 	 "numeric" 	 Ratio of the proportion of Population Aged 65+ to the previous decade.

Details

See pauperism_plu for variables related to the Poor Law Unions.